cookbooks/ai-memory/files/auth-proxy/requirements.txt
aiohttp>=3.9,<4
PyJWT[crypto]>=2.8,<3
